adding middleware and protocols
middleware may be the wrong word, but here we will run corutines
that process messages before they hit a handler by extending the
websockets `.send()` and `.recv()` coroutines. the flow is like
this for `recv` -> recv a message, collect middlewares, pass the
message through the middlewares, return the message. the flow for
`send` -> accept the message to be sent, collect middlewares, pass
the message to send through the middlwares, call `send` with the
processed messages.